week3/name/name.py
def find_names(file_path: str) -> tuple:
    """
    Analyzes a file containing name data and returns a tuple with three elements:
    1. A set of the top three most popular names.
    2. A tuple containing the count and list of names occurring only once.
    3. A tuple containing the initial letter, count,
    and total occurrences of the most common initial letter.
    """
    with open(file_path, 'r', encoding='utf-8') as f:
        f = f.readlines()
        del f[0]
        del f[-1]
        name_list = []
        for i in f:
            i = i.split('\t')
            name_list.append((i[0].strip(),
                              int(i[1].replace('\n', '').replace('(', '').replace(')', ''))))
        for i in range(len(name_list)):
            for j in range(len(name_list)):
                if int(name_list[j][1]) < int(name_list[i][1]):
                    name_list[i], name_list[j] = name_list[j], name_list[i]
        popular_names = set([i[0] for i in name_list[:3]])

        solo = [i[0] for i in name_list if i[1] == 1]
        solo = (len(solo), set(solo))
        letter_dict = {}
        for i in name_list:
            if i[0][0] not in letter_dict:
                letter_dict.update({i[0][0]: 1})
            else:
                letter_dict[i[0][0]] += 1
        for i in letter_dict.items():
            if i[1] == max(letter_dict.values()):
                maxim = i
        k = 0
        for i in name_list:
            if i[0][0] == maxim[0]:
                k += i[1]
        letter_tup = (maxim[0], maxim[1], k)
        return popular_names, solo, letter_tup
